+++ doc/changes.rst
@@ -1,11 +1,11 @@
 Changes
 =======
 
-.. module:: sksparse.cholmod
-
 v0.5.0
 ------
   * Create the :mod:`~sksparse.amd` module, which provides the AMD ordering method.
+  * Create the :mod:`~sksparse.btf` module, which provides the BTF ordering
+    method.
   * Create the :mod:`~sksparse.camd` module, which provides the constrained AMD
     ordering method.
   * Create the :mod:`~sksparse.colamd` module, which provides the COLAMD